+++ blosc/shuffle-neon.h
@@ -28,7 +28,7 @@ BLOSC_NO_EXPORT void shuffle_neon(const int32_t bytesoftype, const int32_t block
 /**
   NEON-accelerated unshuffle routine.
 */
-BLOSC_NO_EXPORT void unshuffle_neon(const size_t bytesoftype, const size_t blocksize,
+BLOSC_NO_EXPORT void unshuffle_neon(const int32_t bytesoftype, const int32_t blocksize,
                                     const uint8_t *_src, uint8_t *_dest);
 
 #ifdef __cplusplus
